requirements.txt
pillow>=8.0.0
requests>=2.23.0
tqdm